Replaced MaxDistanceRetriever with MaxDistanceRetriever2.
