+++ app/src/main/java/org/molgenis/vibe/ontology_processing/MaxDistanceRetriever.java
@@ -2,16 +2,14 @@ package org.molgenis.vibe.ontology_processing;
 
 import org.apache.jena.ontology.OntClass;
 import org.apache.jena.ontology.OntModel;
-import org.apache.jena.util.iterator.ExtendedIterator;
+
 import org.molgenis.vibe.formats.Phenotype;
 import org.molgenis.vibe.formats.PhenotypeNetwork;
 
 import java.util.Collection;
+import java.util.HashSet;
+import java.util.Set;
 
-/**
- * {@link PhenotypesRetriever} implementation that retrieves {@link Phenotype}{@code s} based on a maximum distance from
- * the {@code source}.
- */
 public class MaxDistanceRetriever extends PhenotypesRetriever {
     public MaxDistanceRetriever(OntModel model, Collection<Phenotype> inputPhenotypes, int maxDistance) {
         super(model, inputPhenotypes, maxDistance);
@@ -21,41 +19,52 @@ public class MaxDistanceRetriever extends PhenotypesRetriever {
     public void run() {
         for(Phenotype phenotype:getInputPhenotypes()) {
             PhenotypeNetwork network = new PhenotypeNetwork(phenotype);
-            traverse(retrievePhenotypeFromModel(phenotype), network, 0);
+
+            // previousPhenotypeOCs is an empty Set
+            // currentPhenotypeOCs is a single inputPhenotype
+            Set<OntClass> startOC = new HashSet<>();
+            startOC.add(retrievePhenotypeFromModel(phenotype));
+            traverse(new HashSet<>(), startOC, network, 0);
+
             getPhenotypeNetworkCollection().add(network);
         }
     }
 
     /**
      * Traverses the {@link OntModel}.
-     * @param phenotypeOC current item being traversed
+     * @param previousPhenotypeOCs all {@link Phenotype}{@code s} for {@code distance - 1}
+     * @param currentPhenotypeOCs all {@link Phenotype}{@code s} for the current {@code distance}
      * @param network stores the {@link Phenotype}{@code s} based on traversal
      * @param distance current distance from the {@code network source} (see {@link PhenotypeNetwork#getSource()})
      */
-    private void traverse(OntClass phenotypeOC, PhenotypeNetwork network, int distance) {
-        // Skips certain URIs.
-        if(skippableUri(phenotypeOC)) {
-            return;
+    private void traverse(Set<OntClass> previousPhenotypeOCs, Set<OntClass> currentPhenotypeOCs, PhenotypeNetwork network, int distance) {
+        // For storing phenotypes with a distance + 1
+        Set<OntClass> nextPhenotypeOCs = new HashSet<>();
+
+        for(OntClass phenotypeOC : currentPhenotypeOCs) {
+            // Checks if URI is of a known exception.
+            if (skippableUri(phenotypeOC)) {
+                continue;
+            }
+
+            // Adds current phenotype to the network.
+            addPhenotypeToNetwork(phenotypeOC, network, distance);
+
+            // Looks further if maxDistance is not reached yet.
+            if (distance < getMaxDistance()) {
+                // Adds all the parent and child phenotypes to the "distance + 1" set.
+                nextPhenotypeOCs.addAll(phenotypeOC.listSuperClasses().toSet());
+                nextPhenotypeOCs.addAll(phenotypeOC.listSubClasses().toSet());
+            }
         }
 
-        addPhenotypeToNetwork(phenotypeOC, network, distance);
+        // Removes all "distance - 1" phenotypes from the "distance + 1" set.
+        nextPhenotypeOCs.removeAll(previousPhenotypeOCs);
 
-        // Checks if maximum distance is achieved. If so, returns. If not, continues.
-        if(distance >= getMaxDistance()) {
-            return;
-        } else {
-            // Calculates distance for next recursion step.
-            int nextDistance = distance + 1;
-
-            // Goes through the parents.
-            for (ExtendedIterator<OntClass> it = phenotypeOC.listSuperClasses(); it.hasNext(); ) {
-                traverse(it.next(), network, nextDistance);
-            }
-
-            // Goes through the children.
-            for (ExtendedIterator<OntClass> it = phenotypeOC.listSubClasses(); it.hasNext(); ) {
-                traverse(it.next(), network, nextDistance);
-            }
+        // If there are any phenotypes remaining with "distance + 1", continues.
+        // Note that if the max distance is reached, no phenotypes were ever added to this set.
+        if(nextPhenotypeOCs.size() > 0) {
+            traverse(currentPhenotypeOCs, nextPhenotypeOCs, network, distance + 1);
         }
     }
 }
